% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large-numbers.R
\name{increment}
\alias{increment}
\title{Increment a large number encoded in little endian in constant time}
\usage{
increment(x)
}
\arguments{
\item{x}{raw vector}
}
\description{
Increment a large number encoded in little endian in constant time
}
\examples{
# Increment large number encoded as a raw vector (number is encoded in little endian)
r = as.raw(c(0x00, 0x02))
sodium::increment(r)
}
