% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large-numbers.R
\name{is_zero}
\alias{is_zero}
\title{Test if a large number encoded in little endian is zero (all bytes are zero)}
\usage{
is_zero(x)
}
\arguments{
\item{x}{raw vector}
}
\value{
logical
}
\description{
Test if a large number encoded in little endian is zero (all bytes are zero)
}
\examples{
# Test if
r = as.raw(c(0x00, 0x02))
sodium::is_zero(r)
sodium::is_zero(as.raw(c(0x00, 0x00)))
}
