% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sealing.R
\name{Sealed Box}
\alias{Sealed Box}
\alias{seal_box}
\alias{seal_open}
\title{Sealed Box}
\usage{
seal_box(msg, pubkey)

seal_open(cipher, key)
}
\description{
A \href{http://doc.libsodium.org/public-key_cryptography/sealed_boxes.html}{sealed box}
is sodium's term for basic public key encryption without additional authentication.
}
\examples{
# Generate keypair
key <- keygen()
pubkey <- pubkey(key)

# Encrypt message with pubkey
msg <- serialize(iris, NULL)
cipher <- seal_box(msg, pubkey)

# Decrypt message with private key
out <- seal_open(cipher, key)
stopifnot(identical(out, msg))
}
\references{
http://doc.libsodium.org/public-key_cryptography/sealed_boxes.html
}

