% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hashing.R
\name{hash}
\alias{hash}
\alias{sha256}
\alias{sha512}
\title{Hash functions}
\usage{
sha256(buf, key = NULL)

sha512(buf, key = NULL)
}
\arguments{
\item{buf}{raw vector with data to be hashed}

\item{key}{optional. raw vector with key to be used for HMAC hashing}
}
\description{
Functions to calculate a fixed length hash (digest) of a message. Can also be used
as HMAC by specifying a (sercret) \code{key}. In this case length of the \code{key}
has to match the length of the output of the hash function.
}
\examples{
# Basic hashing
msg <- serialize(iris, NULL)
sha256(msg)
sha512(msg)

# HMAC hash
key256 <- rand_bytes(32)
key512 <- rand_bytes(64)
sha256(msg, key = key256)
sha512(msg, key = key512)
}

