% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signing.R
\name{signing}
\alias{sig}
\alias{sig_keygen}
\alias{sig_pubkey}
\alias{sig_sign}
\alias{sig_verify}
\alias{signing}
\title{Create and Verify Signatures}
\usage{
sig_sign(msg, key)

sig_verify(msg, sig, pubkey)

sig_keygen(seed = rand_bytes(32))

sig_pubkey(key)
}
\description{
Create and verify cryptographics signatures
}
\examples{
# Generate keypair
key <- sig_keygen()
pubkey <- sig_pubkey(key)

# Create signature
msg <- serialize(iris, NULL)
sig <- sig_sign(msg, key)
sig_verify(msg, sig, pubkey)
}

