% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stream.R
\name{streaming}
\alias{aes}
\alias{chacha}
\alias{salsa}
\alias{stream}
\alias{streaming}
\alias{xsalsa}
\title{Stream ciphers}
\usage{
chacha(n, key, nonce)

salsa(n, key, nonce)

xsalsa(n, key, nonce)

aes(n, key, nonce)
}
\arguments{
\item{n}{integer, how many random bytes to generate}

\item{key}{raw vector of size 32 with secret data}

\item{nonce}{non-confidental random data to make the stream unique}
}
\description{
Generate deterministic streams of random data based off a key and nonce.
}
\details{
Random streams form the basis for most cryptographic methods. You usually don't need
to call these methods directly.
}
\examples{
# Very raw encryption
secret <- charToRaw("I like cookies!")
key <- hash(secret)
nonce8 <- rand_bytes(8)
stream1 <- chacha(10000, key, nonce8)
stream2 <- salsa(10000, key, nonce8)

# xsalsa uses bigger nonce
nonce24 <- rand_bytes(24)
stream3 <- xsalsa(10000, key, nonce24)

#' aes uses smaller key
shortkey <- hash(secret, size = 16)
nonce16 <- rand_bytes(16)
stream4 <- aes(10000, shortkey, nonce16)
}

