% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/secret.R
\name{symmetric methods}
\alias{data_decrypt}
\alias{data_encrypt}
\alias{data_tag}
\alias{symmetric methods}
\title{Symmetric Authenticated Encryption}
\usage{
data_encrypt(msg, secret, nonce = rand_bytes(24))

data_decrypt(bin, secret, nonce = attr(cyphertext, "nonce"))

data_tag(msg, secret)
}
\arguments{
\item{msg}{raw vector with message to encrypt or sign}

\item{secret}{raw vector of length 32 with secret key}

\item{nonce}{raw vector of length 24 with non-secret random data}

\item{bin}{raw vector with ciphertext as returned by \code{secret_encrypt}}
}
\description{
Secret key encryption with authentication using a 256 bit key. Mostly useful for
encrypting local data. For secure communication use public-key encryption instead
(\link{secure_send}).
}
\details{
Symmetric encryption uses a secret key to encode and decode a message. This can be
used to encrypt local data on disk, or as a building block for more complex methods.

Because the same \code{secret} is used for both encryption and decryption, symmetric
encryption by itself is impractical for communication. For exchanging secure messages
with other parties, use assymetric (public-key) methods (see \link{encrypt} or
\link{secure_send}).

The \code{nonce} is not confidential but required for decryption, and should be
stored or sent along with the ciphertext. The purpose of the \code{nonce} is to
randomize the cipher to protect gainst re-use attacks. This way you can use one
and the same secret for encrypting multiple messages.

The \link{data_tag} function generates an authenticated hash that can be used to
verify the integrity of piece of data from an earlier generated tag.
}
\examples{
# 256-bit key
secret <- sha256(charToRaw("This is a secret passphrase"))
msg <- serialize(iris, NULL)

# Encrypts with random nonce
cipher <- data_encrypt(msg, secret)
orig <- data_decrypt(cipher, secret)
stopifnot(identical(msg, orig))

# Tag the message with your key (HMAC)
tag <- secret_tag(msg, key)
}
\references{
\url{https://download.libsodium.org/doc/secret-key_cryptography/authenticated_encryption.html}
}

