% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encrypt.R
\name{symmetric methods}
\alias{secret_auth}
\alias{secret_decrypt}
\alias{secret_encrypt}
\alias{secret_verify}
\alias{symmetric methods}
\title{Secret-key authenticated encryption}
\usage{
secret_encrypt(msg, key, nonce = rand_bytes(24))

secret_decrypt(cipher, key, nonce = attr(cipher, "nonce"))

secret_auth(msg, key)

secret_verify(msg, key, tag)
}
\arguments{
\item{msg}{raw vector with message to encrypt or sign}

\item{key}{raw vector of length 32 with secret key}

\item{nonce}{raw vector of length 24 with non-secret random data}

\item{cipher}{raw vector with ciphertext as returned by \code{secret_encrypt}}

\item{tag}{raw vector with a tag as produced by \code{secret_auth}}
}
\description{
Symmetric encryption with authentication using a 256 bit secret key.
See \href{https://download.libsodium.org/doc/secret-key_cryptography/authenticated_encryption.html}{libsodium docs}
for implementation details.
}
\examples{
# 256-bit key
key <- sha256(charToRaw("This is a secret passphrase"))
msg <- serialize(iris, NULL)

# Encrypts with random nonce
cipher <- secret_encrypt(msg, key)
orig <- secret_decrypt(cipher, key)
stopifnot(identical(msg, orig))

# Tag the message with your key (HMAC)
tag <- secret_auth(msg, key)
stopifnot(secret_verify(msg, key, tag))
}
\references{
\url{https://download.libsodium.org/doc/secret-key_cryptography/authenticated_encryption.html}
}

